Add Requires tags for pre*/post* dependencies
This is necessary when elastalert is installed as part of an automated
(kickstart/anaconda-driven) installation.

In some cases, yum/rpm may try to install elastalert before its
dependencies (eg. it tries to run useradd in the preinstallation script
before useradd is actually available, because the shadow-utils package
has not been installed yet).